refactor(server/app): create setup function
