+++ packages/server/src/main.ts
@@ -1,21 +1,10 @@
-import { HttpAdapterHost, NestFactory } from '@nestjs/core';
+import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app/app.module';
-import { ValidationPipe } from '@nestjs/common';
-import cookieParser from 'cookie-parser';
-import { GeneralExceptionsFilter } from './core/filters';
+import { setupApp } from './setup';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  const httpAdapter = app.get(HttpAdapterHost);
-
-  app.useGlobalPipes(
-    new ValidationPipe({
-      whitelist: true,
-      transform: true,
-    }),
-  );
-  app.useGlobalFilters(new GeneralExceptionsFilter(httpAdapter));
-  app.use(cookieParser());
+  await setupApp(app);
   await app.listen(3000);
 }
 bootstrap();
